StringTimes class added
