--- src/main/java/org/PracticeTasks/TestTask.java
@@ -1,6 +1,0 @@
-package org.PracticeTasks;
-
-public class TestTask {
-
-
-}
